rename createProject module as it now only has one function
